Fixed username sorting case-insensitive in admin panel, moved member manager
