+++ members/admin.py
@@ -1,5 +1,10 @@
 from django.contrib import admin
 from django.contrib.auth import admin as auth_admin
+
+
+from django.db.models.functions import Lower
+from members.forms import MemberCreationForm, MemberUpdateForm, SubscriptionPaymentForm
+from members.models import Member, Subscription, SubscriptionPayment, MEMBERSHIP_TYPES
 from django.contrib.auth.models import Permission
 
 from members.forms import (MemberCreationForm, MemberUpdateForm,
@@ -30,7 +35,7 @@ class UserAdmin(auth_admin.UserAdmin):
     list_display = ('username', 'first_name', 'last_name', 'email', 'membership_type', 'is_active', 'is_staff')
     list_filter = ('membership_type', 'is_active', 'groups')
     search_fields = ('first_name', 'last_name', 'email')
-    ordering = ('username', )
+    ordering = [Lower('username'), ]
     readonly_fields = ('last_login',)
     actions = ['activate_user', 'deactivate_user', 'make_ordinary_member', 'make_senior_member']
 
@@ -54,6 +59,8 @@ class UserAdmin(auth_admin.UserAdmin):
         queryset.update(membership_type=SENIOR_MEMBER)
     make_senior_member.short_description = "Sätt medlem till seniormedlem"
 
+    def sorter_username(self, queryset):
+        return Member.objects.all().order_by(Lower('username')).values_list('username', flat=True)
 
 @admin.register(SubscriptionPayment)
 class SubscriptionPaymentAdmin(admin.ModelAdmin):
